Create README.md
Greetings! I'm Abu Reza, a self-motivated and hardworking individual with an extensive interest in research, networking, and communication. Currently, I'm pursuing my honors degree as an undergraduate student majoring in Biotechnology & Genetic Engineering at Islamic University, Bangladesh. I'm so enthusiastic about cancer biology and some of my research projects are underway.
My long-cherished dream is to serve the world as a scientist and for this to happen I tirelessly working so hard every single moment.